Workaround to prevent http end point test failing when soup server can not start
This patch moves the test number 4 up to the end of the test bed, this test has to
do wiht cookie management. Something is not properly managed in there causing the
last reference not be decreased, as a result the soupr server is not being released
casuging next tests fail.
This abnormal behaviour was initially hidden by a work around included in patch
885110f0348439a69e161d1712d0d384403c441b that disabled the checking of the counter
of references at the end of each tests. This error started to come up once the
management of cookies was modified in previous commits.

Change-Id: I642ca416ce6461f4f39de5e4c570f2a73e79cb43