WebSocketTransport: Allow disabling IPv6
In systems where the IPv6 kernel module is not even loaded, the default
behavior of using IPv6 with IPv4 translation will fail because it
depends on having IPv6 enabled in the system.
For these cases, a new setting option allows to control the method used
by Boost to listen for connections, and when `ipv6 = false`, the
IPv4-only method will be used.