debian/default: add G_DEBUG to force crash on GStreamer-CRITICAL
When GLib uses "g_warning" or similar macros, i.e. when GStreamer uses
them to print GStreamer-WARNING or GStreamer-CRITICAL messages, we need
that the variable G_DEBUG is defined to force a breakpoint in them and
be able to obtain either a GDB backtrace or a Kernel core dump that
helps troubleshooting the issue.
Otherwise, the execution continues, and we don't have enough information
in the logs to deduce what went wrong inside GStreamer.