debian minor: use "adduser" instead of "useradd"
Create the kurento user as non-login user. The "--disabled-password"
flag causes the user's shell to be set as "/bin/false" in /etc/passwd

Also don't create the logs directory. That's a job for the initscripts,
that will create the dir appropriately according to the kurento user
specified in /etc/default/kurento-media-server