Delete websocketpp inline sources, use libwebsocketpp instead
websocketpp is the library that provides WebSocket support for Kurento
Media Server. It is a header-only library, so initially it got imported
directly into the project.

Ubuntu 14.04 did not provide this library, but 16.04 onwards do have it,
so we're moving to depend on the package instead of having to deal with
the source code.

This is an advantage because our copy of websocketpp code needs some
adjustments in order to make it compatible with the breaking changes
introduced by OpenSSL 1.1; instead of worrying about this, we'll just
use the Ubuntu's package which already comes with whatever changes are
needed.

Part of https://github.com/Kurento/bugtracker/issues/356
Fixes https://github.com/Kurento/bugtracker/issues/467
Closes https://github.com/Kurento/kurento-media-server/pull/15
Closes https://github.com/Kurento/kurento-media-server/pull/11