debian/init: retry on stop, kill after timeout
When KMS service is stopped, it might happen that the process is so busy
that it doesn't listen to the TERM signal, so it continues working. This
broke the "restart" action, which issues an "stop" immediately followed
by "start", but the second instance of KMS wasn't able to run because
the port 8888 was still in use by the first instance, which was too busy
to realize it should stop.

With --retry=TERM/25/KILL/5 the "stop" action will block until the
server is actually stopped, or otherwise send a KILL signal (forcefully
ends the process) after 25 seconds.

Also align command continuations for start and stop actions.