debian/init: change "Provides" to "kurento-media-server"
https://wiki.debian.org/LSBInitScripts
Normally you should use the script name as boot facility (without .sh if
the file name has such an ending) but one can in the exceptional case
also use the name of the service(s) that the script replaces.

As I interpret this, the Provides section should say "kurento-media-
server" and not ending with a "d" (meaning "daemon").