debian/init: add "coturn" to "Should-Start"
https://wiki.debian.org/LSBInitScripts
Should-Start: boot_facility_1 [boot_facility_2...]
defines the facilities that if present should start before the service
provided by the script. Nevertheless, the script can still start if the
listed facilities are missing. This allows for weak dependencies which
do not cause the service to fail if a facility is not available.

I'm adding this to ensure a clean startup on systems where Coturn is
being used as STUN/TURN server. This line ensures that, if Coturn is
going to be started on boot, KMS will wait for it to be running, thus
avoiding to start the media server before the TURN server.