websocket: document failed keep-alive
When a client tries to use an unknown Session ID, the keep-alive would
fail with a message that was not very informative.

The new one at least tries to guide the user about what is the issue.
Also the comment block talks more about it, for future improvements.

Fixes https://github.com/Kurento/bugtracker/issues/553